Change File Handler to (maybe) save properly
The way ObjectOutputStream works is that it takes an object and saves
everything in it, so if I save it as a GameObject[][] then load it as a
GameObject[][] it will save the objects inside the 2d array. 
Only issue is that I don't believe it saves the positions and anything
within the Grid class thus it has to save a second file which holds just
the grid object.